Added check award limit per user
With this, it will be now possible to set equal to >=, such that if
someone gains for example Reputation without upvotes, the award won't be
missed.

This is also a base for building a better view for the awards, because
it will be possible to only show the number of the same award per user,
instead of printing 4 times the same image.

On the next version I will add more types of auto awards (I was thinking
something like Topic count, days on the forum, upvotes/downvotes given)
and also I will try to beautify the ACP and the all awards page.